// packages/datadog-api-client-v1/models/LogsGrokParserRules.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Set of rules for the grok parser.
*/
export declare class LogsGrokParserRules {
    /**
    * List of match rules for the grok parser, separated by a new line.
    */
    'matchRules': string;
    /**
    * List of support rules for the grok parser, separated by a new line.
    */
    'supportRules'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsGrokParserRules;
    static serialize(data: LogsGrokParserRules): {
        [key: string]: any;
    };
    constructor();
}